+++ pkg/core/types/IStartup.go
